[Refactoring] Move ConvertGuardExprToIfLetExpr to its own file
